backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.3
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.2.5
gunicorn==23.0.0
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
SQLAlchemy==2.0.48
typing_extensions==4.15.0
Werkzeug==3.1.6
zipp==3.23.0
